Update signal-hook to version 0.3.0
The dependency on version v0.1.16 of signal-hook makes this library
impossible to use in a project alongside signal-hook version v0.3.0.
This is caused by an arguably incorrect dependency from signal-hook
v0.1.16 on `~1.2` of signal-hook-registry (I think it should just depend
on `^1.2` instead), however it seemed more beneficial overall to upgrade
this crate's dependency instead.