fix: use signal backend on android
Rustix does not expose pidfd functions for android, so this
commit changes so that the signal backend is used on android.
The pidfd functions were introduced in Linux kernel 5.1, and
epoll integration were introduced in 5.3.

Android version 13, is the first version where all released
version of the Android common kernel is based on Linux kernel
5.4 or higher.
<https://source.android.com/docs/core/architecture/kernel/android-common>

Resolves Issue #79