feat: Use the waiter backend on Windows
In async-process, we have a backend that assumes that child processes
are object that can be `.await`ed on, rather than just being dependent
on signals. At the moment it is only used with Linux and pidfd. Now, it
is used with Windows and the waitable process backend.

At the moment, the backend for `Waitable` in `async-io` is just backed
by a blocking threadpool. However it may also be possible to have it use
IOCP too with little extra overhead. See smol-rs/polling#141 for more
information.

As a side effect, this removes our dependency on `windows-sys`.

Signed-off-by: John Nunley <dev@notgull.net>